refactor: Часть компонента offer вынесена в отдельный комопнент
