+++ src/hooks/use-offer-info.tsx
@@ -2,17 +2,16 @@ import { useEffect } from 'react';
 import { fetchNearbyOffersAction, fetchOfferCommentsAction, getOfferByIdAction } from '../store/api-actions';
 import { useParams } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '.';
-import { FullOffer, OffersData } from '../types/offers';
-import { selectNearbyOffers, selectOffer, selectOffers } from '../store/selectors';
+import { OffersData } from '../types/offers';
+import { selectNearbyOffers, selectOffers } from '../store/selectors';
 import { getNearOffers } from '../utils/offers';
 
-type OfferInfoData = [FullOffer | null, OffersData[]]
+type OfferInfoData = [OffersData[]]
 
 export const useOfferInfo = (): OfferInfoData => {
   const {id} = useParams();
   const dispatch = useAppDispatch();
 
-  const currentOffer = useAppSelector(selectOffer);
   const nearbyOffers = useAppSelector(selectNearbyOffers) as OffersData[];
 
   const allOffers = useAppSelector(selectOffers);
@@ -30,5 +29,5 @@ export const useOfferInfo = (): OfferInfoData => {
     }
   }, [id, dispatch]);
 
-  return [currentOffer, slicedNearbyOffers];
+  return [slicedNearbyOffers];
 };
